requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
requests
python-decouple==3.4
drf-generators
